+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Optional database persistence (via `--persist <path>` flag)
 
 ### Changed
 
@@ -16,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+- API error responses were not fully defined
+- Set `NODE_ENV` to `production` when bundling for NPM
 
 ### Internal