+++ CHANGELOG.md
@@ -8,20 +8,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Optional database persistence (via `--persist <path>` flag)
 
 ### Changed
-- Improve entity stubbing to initialize nullable fields
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
-- API error responses were not fully defined
-- Set `NODE_ENV` to `production` when bundling for NPM
 
 ### Internal
+
+<hr />
+
+## `0.3.0` (2023-Dec-20)
+
+### Added
+- Optional database persistence (via `--persist <path>` flag)
+
+### Changed
+- Improve entity stubbing to initialize nullable fields
+
+### Fixed
+- API error responses were not fully defined
+- Set `NODE_ENV` to `production` when bundling for NPM
 
 <hr />